chore: convert Textarea to Typescript (#1030)
